Move shaded testing dependencies to internal package.
