+++ instrumentation/tapestry-5.4/javaagent/src/test/groovy/TapestryTest.groovy
@@ -10,14 +10,14 @@ import io.opentelemetry.api.trace.SpanKind
 import io.opentelemetry.instrumentation.test.AgentInstrumentationSpecification
 import io.opentelemetry.instrumentation.test.asserts.TraceAssert
 import io.opentelemetry.instrumentation.test.base.HttpServerTestTrait
-import io.opentelemetry.testing.armeria.client.ClientRequestContext
-import io.opentelemetry.testing.armeria.client.DecoratingHttpClientFunction
-import io.opentelemetry.testing.armeria.client.HttpClient
-import io.opentelemetry.testing.armeria.client.WebClient
-import io.opentelemetry.testing.armeria.common.AggregatedHttpResponse
-import io.opentelemetry.testing.armeria.common.HttpHeaderNames
-import io.opentelemetry.testing.armeria.common.HttpRequest
-import io.opentelemetry.testing.armeria.common.HttpResponse
+import io.opentelemetry.testing.internal.armeria.client.ClientRequestContext
+import io.opentelemetry.testing.internal.armeria.client.DecoratingHttpClientFunction
+import io.opentelemetry.testing.internal.armeria.client.HttpClient
+import io.opentelemetry.testing.internal.armeria.client.WebClient
+import io.opentelemetry.testing.internal.armeria.common.AggregatedHttpResponse
+import io.opentelemetry.testing.internal.armeria.common.HttpHeaderNames
+import io.opentelemetry.testing.internal.armeria.common.HttpRequest
+import io.opentelemetry.testing.internal.armeria.common.HttpResponse
 import org.eclipse.jetty.server.Server
 import org.eclipse.jetty.util.resource.Resource
 import org.eclipse.jetty.webapp.WebAppContext
